+++ tinyml-tinyverse/tinyml_tinyverse/references/timeseries_anomalydetection/test_onnx.py
@@ -124,11 +124,11 @@ def get_reconstruction_errors_stats(args):
     logger.info(f"Loading ONNX model: {args.model_path}")
     ort_sess, input_name, output_name = load_onnx_model(args.model_path, args.generic_model)
 
-    errors = torch.tensor([]).to(device, non_blocking=True)
+    errors = torch.tensor([], dtype=torch.float32).to(device, non_blocking=True)
     for _, data, targets in data_loader:
-        data = data.to(device, non_blocking=True).float()
-        targets = targets.to(device, non_blocking=True).long()
-        batch_reconstruction_errors = torch.tensor([]).to(device, non_blocking=True)
+        data = data.float().to(device, non_blocking=True)
+        targets = targets.long().to(device, non_blocking=True)
+        batch_reconstruction_errors = torch.tensor([], dtype=torch.float32).to(device, non_blocking=True)
         for input, target_label in zip(data, targets):
             input = input.unsqueeze(0).cpu().numpy()
             output = torch.tensor(ort_sess.run([output_name], {input_name: input})[0]).to(device)
@@ -172,16 +172,16 @@ def main(gpu, args):
     logger.info(f"Loading ONNX model: {args.model_path}")
     ort_sess, input_name, output_name = load_onnx_model(args.model_path, args.generic_model)
 
-    errors = torch.tensor([]).to(device, non_blocking=True)
-    ground_truth = torch.tensor([]).to(device, non_blocking=True)
+    errors = torch.tensor([], dtype=torch.float32).to(device, non_blocking=True)
+    ground_truth = torch.tensor([], dtype=torch.float32).to(device, non_blocking=True)
 
     for _, data, targets in data_loader:
-        data = data.to(device, non_blocking=True).float()
-        targets = targets.to(device, non_blocking=True).long()
+        data = data.float().to(device, non_blocking=True)
+        targets = targets.long().to(device, non_blocking=True)
         if transform:
             data = transform(data)
-        batch_reconstruction_errors = torch.tensor([]).to(device, non_blocking=True)
-        batch_target_labels = torch.tensor([]).to(device, non_blocking=True)
+        batch_reconstruction_errors = torch.tensor([], dtype=torch.float32).to(device, non_blocking=True)
+        batch_target_labels = torch.tensor([], dtype=torch.float32).to(device, non_blocking=True)
         for input, target_label in zip(data, targets):
             input = input.unsqueeze(0).cpu().numpy()
             output = torch.tensor(ort_sess.run([output_name], {input_name: input})[0]).to(device)
